Experimenting with pushing types inwards.
When we know the type of an expression we should carry that type
inwards. This is particularly important for case expressions where
we want to push the type into the branches. Otherwise we risk losing
a lot of precision if we merge the types of the branches before
checking them with the result type.