Push types to catch errors in clauses
There are now two functions in the type checker for
expressions. One to check an expression against an already specified
type, and one to infer the type. The check function is really good
for catching type errors in case branches and different claueses.

Question: do I need the infer function or is it the same as
checking against any()?