Use the new compatibility relation using subtyping
The type checker should be much lenient now and let a whole
bunch more programs through, not that it's using subtyping.

I've also fixed an issue with the memoization used for the
recursive subtyping. I've stripped the position information
from the types so that it does not get in the way of the
memoization.