Unify opaque types
When encountering a user/remote type that is an opaque type, their
representation is unchanged by normalize/2. This causes a problem when
the same opaque type is compared to itself, because it can be
represented as a remote type or (when extracted from local type or spec
in another module) as a user type.

Unify representation so opaque types are always represented as local
user types in another module.