Support matching on type variables when expecting a list
This patch introduces generation of existentially quantified
variables, which have to be eliminated during constraint solving.
The elimination is not yet implemented.

In this patch the existential type variables are used when
trying to match on a type variable and expecting a list.
If we're matching on 'Var', then we can generate a new type
variable 'EVar', generate the constraint 'Var' <= 'list(EVar)'
and return 'EVar' as the type of the elements of the list.