Support all qualifiers in list comprehension
This patch adds cases for boolean guards and bit string generators in list comprehensions.

For the boolean guards, we (perhaps surprisingly) don't check that the guard returns a boolean. The semantics of Erlang allows for any kind of value in the guard, and only the atom `true` will make the guard evaluate to true. However, it could be considered bad form to return anything else than `true` or `false`. In this case I've decided to be lenient and allow more program, at least until someone can convince me otherwise.