Add support for named funs
It turned out that support for check the type of a named fun was particularly easy. It goes like this:
If we have a named fun like this.
fun F(Args) -> ... end
then we need to give 'F' a type. But during checking we already know that we want the whole expressions to have a type, let's call it 'ResTy'. We can simply assume that 'F' has type 'ResTy' and continue checking the body of the function.