Complete overhaul of function type handling
Function types are a bit fiddly to deal with since they come in many
different varieties. Previously, handling them was intermingled with
checking of clauses.

This patch factors out the handling of function types in a manner
similar to lists and tuples. There's a new helper function
expect_fun_type/1 which does the heavy lifting. As a result, checking
of clauses has been overhauled and is now simpler and less buggy.