Further improvements for handling function types
This patch uses the new machinery for handling function types for
inferring the type of a function call.

This patch also contains more improvements for handling type
checking clauses.

I've some example programs which tests both of these new
improvements.

Ideally, this patch should have been split into two, but it turned out
to be a little cumbersome to disentangle the two improvements