Much improved handling of record patterns
I've introduced machine for matching records akin to the expect_*_type functions already existing for lists, tuples and functions. This patch uses the machinery for record patterns.

Thing to note: the gradualizer now considers tuple() to be a supertype of all record types. This has turned out to be such a prevalent idiom that I've decided that Gradualizer should allow it to pass. Though, I haven't propagated this rule to all the places that need change just yet.