Support MFA with variables for functions
Erlang allows writing "fun mod:Fun/0" which means that the function is
specified with a variable. This patch adds support for this construct.
Many thanks to erszcz on github for bug report #46, which highlighted
this issue among other things.

I probably went overboard in my solution here. I check the type of the
variable and if it is a singleton atom then we actually know what
function the MFA refers to. This means that we can look up the type
and check it. I don't expect the type to be a singleton atom type
very often though.