Support for wildcard assignment in records
Erlang supports the following syntax:

    , _   = bar }

The meaning of this expressions is that all fields except 'apa' will
get the value 'bar'.

I've implemented a fairly strict typing rule for this feature.
Let's assume that the record #rec is declared as follows:

-record(rec, {apa :: atom(), bepa :: atom(), cepa :: integer()}).

This would mean that the above record creation would give the fields
'bepa' and 'cepa' the value 'bar'. This is clearly an error as 'cepa'
has type integer(). Indeed, Gradualizer finds this error by going
through all the unassigned fields and checks their type against the
expression assigned to the wildcard field.

Change-Id: Ic3c218cd6ba62d6ea78746be6078a56a3e715b39