Fix infering type of bool/numeric negation
Arg of not should be sutype of `boolean()` and not the other way
around. Same goes for arg of '-', which should be a subtype of
`number()`.

At the same time negation of literal bool/int and range types is
improved. (Maybe this should only work in case `--infer` is true and
otherwise fall back to eg. -(1) => integer()?)