Fix loading file from within escript
Use erl_prim_loader to get our prelude file using the same mechanism as
the code server.

An exception is when the file cover compiled. Then we load it from the
source code, assuming it is in the same directory as the source of the
current module. Thus, cover compiled prelude when running as an escript
archive doesn't work but we can live with that.