Bugfix: Allow non-empty concrete lists
We used to report an error if we had a concrete list like '[1]' and
it should have type 'nonempty_list()'. Why? Because the list is
represented as a cons followed by a nil, and we require the tail of
the cons to have the same type as the whole cons. But this patch
weakens the demand on the type of the tail of the cons so that
it is always allowed to be the empty list.