Fix handling undefined records and record fields
In practice Gradualizer is mostly run on code that compiles without
errors. (i.e. there should be no undefined records or record fields)
However we don't make such assumptions and if an undefined record or
record field is encountered it is still reported in a user friendly
way. (Let's note though that Gradualizer won't make extra efforts to
detect all undefined records or record fields if they are not needed for
type checking)