Print column numbers in errors if available
This will only work if the abstract forms are aquired from an erlang
source file. Debug info from a beam file won't contain column numbers.

Many places in `handle_type_error` assume the annotation is a single
integer representing the line number, whereas it can be a location
two-tuple now. But the printouts will be fixed gradually as we unify
errors.