New specs for functions in the lists module
Some of the specs in the lists module in otp don't work so well with the gradualizer.
This patch gives new specs which overrides the otp specs.

Dialyzer crashes when given the gradualizer_prelude. It is probably
confused by multiple specs for the same function. Therefore, we're
hiding that module from the dialyzer.