Fix record compatibility check
Have to remove position before comparing types of two record fields.
It can easily happen that two modules include the same headerfile, but
with different line numbers.

Even so when self-gradualizing in typechecker there was a fully
qualified self-call (?MODULE:fn) so the spec was looked up by
gradualizer_db which fetched it probably from the beam file while the
analysis was run on source. This resulted in different positions even for
the same module as the AST from source has line numbers but the AST from
debug_info doesn't.

This fixes:
src/typechecker.erl: The variable ParseData is expected to have type #parsedata{} but it has type #parsedata{}