Enable exhaustiveness checking
This patch enables exhaustiveness checking for the gradualizer.

Not all types can be checked for exhaustiveness and we also don't
handle types in guards at the moment. But it does work for a couple of
limited cases.

This patch reuses the type refinement mechanism and checks that if we
the final refined types after having gone through all the clauses is
none(), then the clauses are exhaustive.

Unfortunately, the type refinement machinery is a little too coarse
since we must be able to refine to a sensible erlang type at each
intermediate step and that hinders us from capturing certain patterns
such as various forms of patterns on lists. Eventually, we will have
to implement a completely different machinery for exhaustiveness
checking to get something more precise, but this will do for now.

Implementing exhaustiveness checking revealed a bunch of imprecisions
in the type refinement code and I've fixed them in this patch.

A few test cases needed to be updated but I'm surprised that there
wasn't more of them.