Normalise the case for multiple variables
We merge the information we get from one guard sequence into the types
given by the fun specs, so that a sequence guard has information about
all the function's variables.

When calculating the least upper bound of all guard sequences, if
there's any sequence that gives a variable the type any, then we don't
really know the type of the variable and we should therefore keep it as
just any.