Enable overriding cli options with attributes
The PR #212 introduced the `-gradualizer()` attribute, which allows
setting options for gradualizer in a module. I would have expected
these directives to override options on the command line, which is the
behaviour in other compilers and in dialyzer. This patch makes sure
that the attributes override cli options properly.

It also changes slightly how some options are treated internally.