Change cli options to use underscores to match the erlang api
This patch implements the suggestion in #213 to harmonize the way
options are written. It used to be that we have e.g. `--no-infer` in
the command line but `no_infer` when calling gradualizer from the
api. Now we will use `--no_infer` on the command line.

This spelling of the options matches the convention of dialyzer.