Fancy error messages
The expression is printed highlighted in its context. This
works only if no custom pretty-printing function is used. If
so, it falls back to the non-fancy printing automatically.

If the source code is available, it is used. Otherwise, it is
recreated from the AST by pretty-printing (Erlang only).

Colors are used if a TTY can be detected, using io:columns().
This doesn't work in escripts, so --color has to be specified
explicitly for now.

Fancy error messages can be printed without color too. To turn
off fancy messages completely, use --no_fancy.