Fix variable capture in 'fun'
There's been a long-standing TODO about how variable capture
is dealt with in certain situations when adding types to
variables in clauses. This is handled differently between
'fun's and other types of constructs which have clauses
such as 'case' and 'receive'.

This patch passes around information to make sure that when
variables are handled in clauses, it is possible to know whether
variables should be bound or captured. It's a bigish diff but
it's mostly pretty boring.

Fixes #138