A first stab at typechecking automatically from the repl
When the module 'user_default' is loaded it will override
the compile command in the repl so that files are typechecked
before compiled.

There are plenty of ways this implementation can be improved.
Right now the command only accepts strings as filepaths pointing
to source code files.
Secondly, it parses the file and passes it on to gradualizer.
After typechecking the absform is completely discarded and the
compiler starts parsing from scratch.