Support for better location information with constraints
Constraints don't have good error reporting as of now. This commit
adds support for keeping track of locations so that we can report it
when we get a type error. It doesn't solve the whole problem though,
because we need type information in types as well in order to pass the
information to the constraints