Remove most warnings when compiling the test suite
Warnings can be useful even when compiling the test suite so I've
taken some time to remove as many warnings as I can.

I've replaced some `export` pragmas with
`-compile([export_all,nowarn_export_all])`. I don't think it makes
sense to maintain an export list in the test suite, as there isn't
any abstraction defined in those modules that is in need of protection
by not exporting all functions.

There are a few types of warnings that I haven't been able to
remove, and if someone knows how, I'd be interested in learning:

* "this clause cannot match because a previous clause at line 6 always
  matches"
  In some tests we explicitly want to test overlapping matches, e.g.
  type refinement, so this warning is not helpful.
* "this expression will fail with a 'badarith' exception"
  Similar to above, we want to test that typechecking fails when
  arithmetic expressions are given the wrong type of arguments.