Make sure not to recurse forever in typechecker:normalize()
Given a recursive type, typechecker:normalize() can recurse forever if
we don't take care to stop the recursion. This patch adds a set to
keep track of the types that `normalize` has unfolded so far and stops
unfolding if asked to normalize a type that has already been unfolded.

Ilya identified the problem with recursives types in #322 and offered
a nice example program to demonstrate the problem.

Fixes #322 although the program is still not accepted as it should be
so I'm adding it to the known problems.