Clean up hack in `glb` and `add_any_types_pat`
In `add_any_types_pat` we called `glb` with one of the arguments
being `any()`. But that always returns `any()`! So we can simplify
that whole call away.
Furthermore, the call to `glb` used an `empty_tenv` which is going
to crash since `glb` needs the module name from the `tenv` but
`empty_tenv` doesn't set the module name.
So now, instead of hacking around the fact that `glb` might not get
a proper module name, we will simply always fail. This will help
us catch bugs like these faster in the future.

This issue was highlighted by @erszcz in #333. Thanks!