Improve typechecking maps
This is an incremental improvement to typechecking maps.
It's not final, in the sense that there are still known cases that are not typechecked properly.

In short, this should work properly on maps with non-overlapping keys, such as maps with atom keys
and no "wildcard" association (any() => any()). This means Elixir structs should be handled properly.

In a bit more detail, this commit brings the following changes:

- glb() on maps is still short-circuited to return none() if a valid result cannot be computed,
  but it tries to return sensible results in more cases than before.
  If keys of any of the map params overlap we return none().
  If they do not, we compute the GLB.
  This means we return a GLB in more situations than before, but still decline to compute it
  if negation types are required to do it properly.
- map type inference of map creation expressions
- subtype (or actually compat_ty) logic changes to follow the wiki
- handling of map unions in update_map_type
- map normalisation, so they're easier to handle
- fix to a name clash between test/should_pass/map_expr.erl
  and test/known_problems/should_fail/map_expr.erl
- new map test cases as well as some solved known_problems