Pass union_size_limit in env()
Prior to this commit, the Env got passed around, but was not really used by normalize().
This meant that even an invalid Env didn't trigger any bugs.
Now that we're using the Env to get the union size limit, it started triggering bugs where the env
was not correctly adjusted to env() from tenv().