Fix glb's use of normalize causing recursive types to grow indefinitely
Let's consider the following type Ty1:

    t2(#{#{t1() := <<_:$Z, _:_*$q>>, tuple() => fun()} =>
                       #r3{f1 :: t2(t1())}})

And type env:

    -type t1() :: {}.
    -type t2(A1) :: #{#{t1() := <<_:$Z, _:_*$q>>, tuple() => fun()} =>
                    #r3{f1 :: t2(t1())}} | A1.

Now, let's consider the following:

    Ty1Norm1 = typechecker:normalize(typelib:remove_pos(Ty1), Env).
    Ty1Norm2 = typechecker:normalize(typechecker:normalize(typelib:remove_pos(Ty1), Env), Env).

	39> io:format("~ts\n", [typelib:pp_type(Ty1Norm1)]).
	#{#{{} := <<_:$Z, _:_*$q>>, tuple() => fun((...) -> any())} =>
					#r3{f1 ::
							t2(#{#{t1() := <<_:$Z, _:_*$q>>,
								   tuple() => fun()} =>
									 #r3{f1 :: t2(t1())}})}}

    41> io:format("~ts\n", [typelib:pp_type(Ty1Norm2)]).
    #{#{{} := <<_:$Z, _:_*$q>>, tuple() => fun((...) -> any())} =>
                    #r3{f1 ::
                            #{#{{} := <<_:$Z, _:_*$q>>,
                                tuple() => fun((...) -> any())} =>
                                  #r3{f1 ::
                                          t2(#{#{t1() := <<_:$Z, _:_*$q>>,
                                                 tuple() => fun()} =>
                                                   #r3{f1 :: t2(t1())}})}}}}

What happens above?
When a type is normalized, the topmost occurences of user types in its definition are expanded.
If the user types are recursive, a subsequent normalization leads to their further expansion.
This means, that normalize^1(Ty1) is NOT equal to normalize^2(Ty1), where f^2(x) := f(f(x)).
In other words, subsequent normalizations of such a type make the type grow indefinitely.

We have to keep this behaviour in mind when doing recursive computations on types,
to avoid infinite loops when trying to process the ever-growing type.
In other words, normalize just once, then recurse down the resulting type.