Recover position information for undef/not_exported type errors
This catches undef/not_exported type errors on a level where position information is available
from function specs and rethrows with this information supplied in the exception.
In order to source this position information from specs it has to be stored in #env.fenv,
so it is NOT removed from specs in `create_fenv()` anymore - this has to be done as required after
fetching specs from #env.fenv.