Disable exhaustiveness checking on nonexhaustive list pattern inference
Apart from the change above, this changes also tweaks the pp_intersection test.
Before this commit, the result of this test was:

$ ./bin/gradualizer test/should_fail/pp_intersection.erl
test/should_fail/pp_intersection.erl: The clause on line 12 at column 1 cannot be reached
test/should_fail/pp_intersection.erl: None of the types of the function foo on line 15 at column 11 matches the call site. Here's the types of the function:
fun((integer()) -> float())
fun((Float) -> integer() when Float :: float())

However, I don't think it was correct, since there's really no reason for the code not to reach
the clause on line 12 (2nd clause of function foo) - it seems the warning was a false positive.

Currently, with the test fix in this commit, there's only one error reported, the latter one:

test/should_fail/pp_intersection.erl: None of the types of the function foo on line 15 at column 11 matches the call site. Here's the types of the function:
fun((integer()) -> float())
fun((Float) -> integer() when Float :: float())