Add known problem for unsupported maybe expression
The syntax for the new experimental maybe expression is currently not
supported. Highlight and document this fact by putting it in the known
problems test directory.

Avoid compiling with support for the maybe feature outside of test
profiles to avoid causing issues in people depending on gradualizer
trying to load modules.