Make maybe_expr known problem work both on pre- and post-25 Erlang/OTP versions
Due to feature gating we cannot just rely on an `ok | nok` response on a given test file.
Instead, we make gen_known_problem_should_pass() expect a specific number of errors.
For known_problems/should_pass, we expect a number of errors equal to the number of exported
functions, i.e. each exported function is a single known problem.
This works both if the gated feature is enabled or not.