Remove call to rewrite_map_assocs_to_exacts
In the CheckAssoc lambda function inside add_type_pat
for maps, we already create all associations with
map_field_exact, so there is no need to rewrite it.
This is probably a relic of some refactoring.

I am removing the function rewrite_map_assocs_to_exacts
as this was the only place where it was used.