add_type_pat for maps: Add non-matched keys to the pattern type
This handles cases when you don't match on all of the fields
of the map. For instance, if you have the type #{x := a, y := b}
and the pattern #{x := a}. This pattern matches the value #{x := a}
but also #{x := a, y := b}. The resulting pattern type thus has
to be #{x := a, y => b}.

It also handles cases when you match only on part of the key type.
For instance, if you have the type #{x|y := a} and the pattern
is #{x := a}, it matches both #{x := a} and #{x := a, y := a},
and thus the pattern should be #{x := a, y => a}.

Unfortunately, this commit fixes stuff but it also breaks other
stuff. If you have a pattern with duplicate keys
(like #{x := a, x := a}), Gradualizer will now report a key
error, although this pattern is completely okay (given that the
type is for example #{x := a}). However, this is a really rare
use of pattern matching, and can always be rewritten (by just
dedupliting the keys in the pattern). Still, it should be fixed
in the future.