Rewrite the refine_ty algorithm for maps
The current algorithm was too simplistic and didn't work in many cases.
Most of the newly added golden tests are failing (producing wrong results)
on master.

Together with the rewrite of add_pats_types for maps, it should work
correctly for most cases. The only known problem is that maps with
non-singleton required keys are not checked for exhaustiveness. But
this should not cause much trouble because such maps aren't often
used in practice.