Add `apps` option to rebar3 commands
This option allows to specify a list of apps where to run the type
check. The list can be specify when running the command or in
`rebar.config`.

This option can be really handy when trying to integrate Gradualizer on
big umbrella project. If the project contains a lot of errors, it is
possible to enable the type check app by app and fix the errors
incrementally.