Merge intersection function types using inner unions
Consider a function spec:
    -spec negate(bool()) -> bool();
                (integer()) -> integer().

Sometimes, we want to make a single function type out of these two clauses.
Using an outer union like fun((bool()) -> bool()) | fun((integer()) -> integer())
leads to a loss of precision, because this type says that we either have
a function that can take a bool, or we have a function that can take an integer
(and we do not know which case it is). This means that we cannot apply a function
of this type to anything unless we assert type or something like that.

But we know for sure that it can take both, bools and integers! Therefore, using
an inner union like fun((bool() | integer()) -> bool() | integer()) seems better
to me as this way we know that it can take both.

Of course, an ideal solution would be to use intersection types like
fun((bool()) -> bool()) & fun((integer()) -> integer()), but Erlang doesn't have
them in its syntax and so don't we in Gradualizer.