Document a known problem related to union function types
Union function types (`fun_ty_union`) do not seem to work. These simple
examples should be perfectly valid. I think it throws that error because
`expect_fun_type/3` inputs a list of function types (i.e., a clause),
containing only a single element (the union of two function types).
This leads to expecting an intersection function type later on
(which is wrong).

Note that these union function types aren't that much useful in practice
as when you want to call them, all your argument types must be subtypes
of parameter types of all the possible function types. That is, when
you have a function of the type
    fun((integer()) -> integer()) | fun((atom()) -> atom())
then you can only call it with `none()` as we don't know whether the
function will accept integers or bools and no other type is a subtype
of `integer()` and `atom()`.