Use Local Type Inference approach to typecheck polymorphic calls
We use the Local Type Argument Synthesis algorithm from
https://www.cis.upenn.edu/~bcpierce/papers/lti-toplas.pdf
(a bit extended for our use case).

This replaces the global constraint generation and constraint
solver. It should result in a much more precise type error
messages (pinpointing the exact location, not just the
function) and more streamlined code (refactor will be made
in a subsequent commit).

For more details, see the original paper or
https://github.com/josefs/Gradualizer/wiki/Polymorphism#typechecking-polymorphic-calls