Remove constraints from where they are not needed
Flexible type variables may now appear only when typechecking a polymorphic call,
in which case they are only in the type of that polymorphic function. No other
expression may of be a type that contains a flexible type variable. Constraints
are thus needed only in the subtype_with_constraints/3 function that gets called
only from type_check_poly_call/4. This means that nothing but the subtyping
functions and the type_check_poly_call/4 function needs to deal with constraints.
Therefore, I am removing constraints from all the other places.

I believe this is a great simplification of the code. It also trivially solves
all the TODOs like "Don't drop the constraints".

In such a huge refactoring, Gradualizer has proved to be an invaluable help.
Via self-gradualization, it discovered numerous errors that I didn't notice.
Hopefully, there are no undetected errors left. All tests are passing, the
self-gradualization does not crash and 100_000 proptests have also passed.