any_type/4: return constraints if exactly one supertype matches
This quite a trivial change enables us to typecheck polymorphic
functions like

    -spec generic_hd([A,...] | {A, any()}) -> A.

or functions from Elixir's Enum module that use protocol
polymorphism in addition to the parametric polymorphism:

    -spec enum_map([A] | map(), fun ((A) -> B)) -> [B].

which is really handy for Gradient and other Elixir frontends!

Support for intersection polymorphic functions like

    -spec generic_hd([A,...]) -> A;
                    ({A, any()}) -> A.

is of course still missing. But at least some of these can now
be typechecked when using a spec with the unions.