Remove the --infer option
This removes the --infer CLI and #env{} option.
The Erlang VM does dynamic type checking based on type tags embedded in values.
This means that it's safe to infer type information from literals and operators
according to the same rules that the VM enforces at runtime.
Actually, these are one of the few completely credible sources of information,
unlike type specs, in which there's room for human error.
Removing the option along with the branching in the typechecker should somewhat
minimise the amount of code to maintain.